github: Detect merge mode using change trees
The previous algorithm to detect how a PR was merged compared the
commit trees to detect if a series of commits in a PR lined up in
a branch.

Unfortunately, the tree hashes are modified if commits are merged
between the PR merge and the cherrypick.

This commit adds a new algorithm to checksum changes into a custom
`ChangeTree` which hashes (sha256) all the changes in a commit, but
leaving out files not modified (git trees include all changes in the
repo).

Signed-off-by: Adolfo García Veytia (Puerco) <adolfo.garcia@mattermost.com>