run: StagingPath function
This commit adds a new function to the Run object to generate
predictable paths to stage artifacts according to the hashes
identifying its naterials.

The algorithm takes into account the build point and the
material URL and hashes and performs the following algo:

	// The algorithm to determine the staging path is:
	// 1. Sort the materials by URI
	// 2. Concat: buildpoint + (materials.URL[n]+materials.Sha[n])
	// 2a: Sha should be the first sha found using: this order: sha1 sha256 sha512 (else fail)
	// 3. Hash the whole string sha256

A small test to check the expected outputs is included in the
new run unit tests file.

Signed-off-by: Adolfo García Veytia (Puerco) <adolfo.garcia@mattermost.com>